[BUGFIX] Pagination wrong in IndexedSearch
Fix variable type issues in IndexedSearch pagination.
(classic and extbase VH)

This will ensure the CSS classes are correct again for current page and
next/prev page.

Resolves: #63350
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/34646
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Markus Klein <klein.t3@reelworx.at>

+++ typo3/sysext/indexed_search/Classes/Controller/SearchFormController.php
@@ -1495,7 +1495,7 @@ class SearchFormController extends \TYPO3\CMS\Frontend\Plugin\AbstractPlugin {
 		$pointer = (int)$this->piVars['pointer'];
 		$count = (int)$this->internal['res_count'];
 		$results_at_a_time = \TYPO3\CMS\Core\Utility\MathUtility::forceIntegerInRange($this->internal['results_at_a_time'], 1, 1000);
-		$pageCount = ceil($count / $results_at_a_time);
+		$pageCount = (int)ceil($count / $results_at_a_time);
 
 		$links = array();
 		// only show the result browser if more than one page is needed
